Restore implementation of persistent state
The implementation is race free. There is, however, a known polling loop in the even that the daemon has already decided to go down
at the time a new client is trying to connect, but has not, yet, released the lock. The client will busy-loop until the lock is
released, at which point the client will spawn a new daemon.