Document missing handling of posix_acl_access
system.posix_acl_access is used by "cp -a" even in places "chmod" would
do just as fine. This causes the file_lie database and the disk to
disagree on the file's permissions, which triggers the fail-safe
mechanism and erases the information from the database, reverting the
file to its real stat.